hw2/115701571_hw2/pyproject.toml
[project]
name = "115701571_hw2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "glfw>=2.7.0",
    "pyglm>=2.7.2",
    "pyopengl>=3.1.7",
]
